refactor(app): clean up mission component structure
